requirements.txt
pybit
websocket-client
pandas
numpy
ta
requests